+++ docs/changelog.md
@@ -1,6 +1,7 @@
 ## 0.2.x (Current)
 - Graph force layout support
 - Support SCgStruct
+- Support link content in SCgStruct
 
 ## 0.1.x